Make code work with photographed images as well
Images that were barely rotated were problematic as the line of pixels
containing the first corner would go gradually from black to white and
cross the threshold somewhere that was not the corner. So when the
first corner is on the right side, it is checked that the three next
pixels are below the threshold to make sure we're not in the middle of
the line somewhere.